refactor: separa camada controller e service do auth
@eluvya @patrciatersaiol

64
