+++ src/controllers/auth/index.ts
@@ -1,36 +1,24 @@
-const jwt = require('jsonwebtoken')
+import { HttpResponseHandler } from '@controllers/HttpResponseHandler'
+import { createAccessToken, validateAccessToken } from 'src/service/auth/authService'
+
+const responseHandler = new HttpResponseHandler()
 
 export const generateAccessToken = (request, response) => {
-  const { EMAIL_ADMIN, PASSWORD_ADMIN, NAME_ADMIN, SECRET } = process.env
   const emailUser = request.body.email
   const passwordUser = request.body.password
 
-  if (emailUser !== EMAIL_ADMIN || passwordUser !== PASSWORD_ADMIN) {
-    return response.sendStatus(401)
+  try {
+    const tokenPayload = createAccessToken(emailUser, passwordUser)
+    return response.json(tokenPayload)
+  } catch (error) {
+    return responseHandler.createErrorResponse(error, response)
   }
-
-  const payload = { name: NAME_ADMIN, email: EMAIL_ADMIN }
-  const accessToken = jwt.sign(payload, SECRET)
-
-  return response.json({
-    auth: true,
-    accessToken,
-    user: payload
-  })
 }
 
 export const verifyAccessToken = (request, response, next) => {
-  const { SECRET } = process.env
   const authHeaders = request.headers.authorization
-  const accessToken = authHeaders && authHeaders.split(' ')[1]
 
-  const isAuthenticated = jwt.verify(
-    accessToken,
-    SECRET,
-    (err) => {
-      return !err
-    }
-  )
+  const isAuthenticated = validateAccessToken(authHeaders)
 
   if (isAuthenticated) {
     next()
