+++ src/service/HttpError.ts
@@ -1,6 +1,7 @@
 export enum HttpStatusCode {
   OK = 200,
   BAD_REQUEST = 400,
+  UNAUTHORIZED = 401,
   NOT_FOUND = 404,
   INTERNAL_SERVER = 500,
 }
